+++ R/FoReco-package.R
@@ -25,11 +25,12 @@
 #' }
 #'
 #' @references
-#' Di Fonzo, T., Girolimetto, D. (2020), Cross-Temporal Forecast Reconciliation:
-#' Optimal Combination Method and Heuristic Alternatives, Department of Statistical
-#' Sciences, University of Padua, \href{https://arxiv.org/abs/2006.08570}{arXiv:2006.08570}.
+#' Di Fonzo, T., and Girolimetto, D. (2021), Cross-temporal forecast reconciliation:
+#' Optimal combination method and heuristic alternatives, \emph{International Journal
+#' of Forecasting}, in press \doi{10.1016/j.ijforecast.2021.08.004}.
 #'
-#' Di Fonzo, T., Girolimetto, D. (2021), \emph{Forecast combination based forecast reconciliation: insights and extensions} (mimeo).
+#' Di Fonzo, T., Girolimetto, D. (2021), Forecast combination based forecast reconciliation:
+#' insights and extensions, \href{https://arxiv.org/abs/2106.05653}{arXiv:2006.08570}.
 #'
 #' @import mathjaxr
 #' @docType package