+++ R/commat.R
@@ -2,7 +2,7 @@
 #'
 #' @description
 #' \loadmathjax
-#' This function returns the (\mjseqn{(r c) \times (r c)})
+#' This function returns the (\mjseqn{r c \times r c})
 #' commutation matrix \mjseqn{\mathbf{P}} such that
 #' \mjsdeqn{\mathbf{P} \mbox{vec}(\mathbf{Y}) = \mbox{vec}(\mathbf{Y}'),}
 #' where \mjseqn{\mathbf{Y}} is a (\mjseqn{r \times c}) matrix.
@@ -10,7 +10,7 @@
 #' @param r Number of rows of \mjseqn{\mathbf{Y}}.
 #' @param c Number of columns of \mjseqn{\mathbf{Y}}.
 #'
-#' @return A sparse (\mjseqn{(r c) \times (r c)}) matrix, \mjseqn{\mathbf{P}}.
+#' @return A sparse (\mjseqn{r c \times r c}) matrix, \mjseqn{\mathbf{P}}.
 #'
 #' @references Magnus, J.R., Neudecker, H. (2019), Matrix Differential Calculus
 #' with Applications in Statistics and Econometrics, third edition, New York,