+++ R/octrec.R
@@ -29,11 +29,13 @@
 #' of variables:
 #' \itemize{
 #'   \item \bold{ols} (Identity);
-#'   \item \bold{struc} (Cross-temporal summing matrix);
-#'   \item \bold{wlsh} (Hierarchy variances matrix);
-#'   \item \bold{wlsv} (Series variances matrix);
+#'   \item \bold{struc} (Cross-temporal structural variances);
+#'   \item \bold{cs_struc} (Cross-sectional structural variances and temporal independence);
+#'   \item \bold{t_struc} (Cross-sectional independence and temporal structural variances);
+#'   \item \bold{wlsh} (Hierarchy variances);
+#'   \item \bold{wlsv} (Series variances);
 #'   \item \bold{bdshr} (Shrunk cross-covariance matrix, cross-sectional framework);
-#'   \item \bold{bdsam}  (Sample cross-covariance matrix, cross-sectional framework);
+#'   \item \bold{bdsam} (Sample cross-covariance matrix, cross-sectional framework);
 #'   \item \bold{acov} (Series auto-covariance matrix);
 #'   \item \bold{Sshr} (Series shrunk cross-covariance matrix);
 #'   \item \bold{Ssam} (Series cross-covariance matrix);
@@ -224,7 +226,7 @@
 #' \item{\code{varf} (\code{type="direct"})}{(\mjseqn{n \times (k^\ast + m)}) reconciled forecasts variance matrix for \mjseqn{h=1}, \mjseqn{\mbox{diag}(\mathbf{MW}}).}
 #' \item{\code{M} (\code{type="direct"})}{Projection matrix (projection approach).}
 #' \item{\code{G} (\code{type="S"} and \code{type="direct"})}{Projection matrix (structural approach, \mjseqn{\mathbf{M}=\mathbf{S}\mathbf{G}}).}
-#' \item{\code{S} (\code{type="S"} and \code{type="direct"})}{Cross-temporal summing matrix (\mjseqn{\widetilde{\textbf{S}}\mbox{vec}(\widehat{\textbf{Y}}')} representation).}
+#' \item{\code{S} (\code{type="S"} and \code{type="direct"})}{Cross-temporal summing matrix (\mjseqn{\widetilde{\textbf{F}}\mbox{vec}(\widehat{\textbf{Y}}')} representation).}
 #' \item{\code{info} (\code{type="osqp"})}{matrix with some useful indicators (columns)
 #' for each forecast horizon \mjseqn{h} (rows): run time (\code{run_time}), number of iteration,
 #' norm of primal residual (\code{pri_res}), status of osqp's solution (\code{status}) and
@@ -234,9 +236,9 @@
 #' Byron, R.P. (1978), The estimation of large social accounts matrices,
 #' \emph{Journal of the Royal Statistical Society A}, 141, 3, 359-367.
 #'
-#' Di Fonzo, T., Girolimetto, D. (2020), Cross-Temporal Forecast Reconciliation:
-#' Optimal Combination Method and Heuristic Alternatives, Department of Statistical
-#' Sciences, University of Padua, \href{https://arxiv.org/abs/2006.08570}{arXiv:2006.08570}.
+#' Di Fonzo, T., and Girolimetto, D. (2021), Cross-temporal forecast reconciliation:
+#' Optimal combination method and heuristic alternatives, \emph{International Journal
+#' of Forecasting}, in press.
 #'
 #' \enc{Schäfer}{Schafer}, J.L., Opgen-Rhein, R., Zuber, V., Ahdesmaki, M.,
 #' Duarte Silva, A.P., Strimmer, K. (2017), \emph{Package `corpcor'}, R
@@ -294,7 +296,8 @@ octrec.default <- function(basef, m, C, comb, res, Ut, nb, mse = TRUE,
 
   comb <- match.arg(comb, c(
     "ols", "struc", "sam", "wlsv", "wlsh", "shr",
-    "acov", "Ssam", "Sshr", "bdshr", "bdsam", "w", "omega"))
+    "acov", "Ssam", "Sshr", "bdshr", "bdsam", "w", "omega",
+    "cs_struc", "t_struc"))
 
   type <- match.arg(type, c("M", "S"))
   keep <- match.arg(keep, c("list", "recf"))
@@ -321,7 +324,7 @@ octrec.default <- function(basef, m, C, comb, res, Ut, nb, mse = TRUE,
   na <- ctf$hts$na
   nb <- ctf$hts$nb
   #C <- ctf$hts$C
-  #S <- ctf$hts$S
+  Scs <- ctf$hts$S
   #Ut <- ctf$hts$Ut
 
   kset <- ctf$thf$kset
@@ -329,6 +332,7 @@ octrec.default <- function(basef, m, C, comb, res, Ut, nb, mse = TRUE,
   p <- ctf$thf$p
   kt <- ctf$thf$kt
   ks <- ctf$thf$ks
+  R <- ctf$thf$R
 
   # matrix
   #Zt <- tools$Zt
@@ -338,7 +342,7 @@ octrec.default <- function(basef, m, C, comb, res, Ut, nb, mse = TRUE,
   }
 
   Ht <- ctf$ctf$Ht
-  S <- ctf$ctf$Stilde
+  S <- ctf$ctf$Fmat
 
   # Base forecast
   if (NCOL(basef) %% kt != 0) {
@@ -353,15 +357,15 @@ octrec.default <- function(basef, m, C, comb, res, Ut, nb, mse = TRUE,
 
   if(is.null(S)){
     if (type == "S") {
-      stop("Type = S needs the cross-sectional C matrix. \nPlease consider using ut2c() to get the right C when working with Ut.", call. = FALSE)
+      stop("Type = S needs the cross-sectional C matrix. \nPlease consider using gecoma() to get the right C when working with Ut.", call. = FALSE)
     }
 
     if(nn_type == "sntz"){
-      stop("nn_type = sntz needs the cross-sectional C matrix.\nPlease consider using ut2c() to get the right C when working with Ut.", call. = FALSE)
+      stop("nn_type = sntz needs the cross-sectional C matrix.\nPlease consider using gecoma() to get the right C when working with Ut.", call. = FALSE)
     }
 
-    if (comb == "struc") {
-      stop("Param comb equal struc needs the cross-sectional C matrix.\nPlease consider using ut2c() to get the right C when working with Ut.", call. = FALSE)
+    if (comb == "struc" | comb == "cs_struc") {
+      stop("comb = ", comb, " needs the cross-sectional C matrix.\nPlease consider using gecoma() to get the right C when working with Ut.", call. = FALSE)
     }
   }
 
@@ -412,7 +416,7 @@ octrec.default <- function(basef, m, C, comb, res, Ut, nb, mse = TRUE,
   }
 
   if (corpcor) {
-    shr_mod <- function(x, ...) corpcor::cov.shrink(x, verbose = FALSE, ...)
+    shr_mod <- function(x, ...) Matrix(unclass(corpcor::cov.shrink(x, verbose = FALSE, lambda.var=0, ...)))
   } else {
     shr_mod <- function(x, ...) shrink_estim(x, minT = mse)[[1]]
   }
@@ -423,6 +427,12 @@ octrec.default <- function(basef, m, C, comb, res, Ut, nb, mse = TRUE,
          },
          struc = {
            Omega <- .sparseDiagonal(x = rowSums(S))
+         },
+         "cs_struc" = {
+           Omega <- .sparseDiagonal(x = rep(rowSums(Scs), each = kt))
+         },
+         "t_struc" = {
+           Omega <- .sparseDiagonal(x = rep(rowSums(R), n))
          },
          sam = {
            Omega <- cov_mod(E)
@@ -483,7 +493,7 @@ octrec.default <- function(basef, m, C, comb, res, Ut, nb, mse = TRUE,
   if(!is.null(v)){
     keep <- "recf"
     rec_sol <- recoV(
-      basef = Ybase, W = Omega, Ht = Ht, sol = sol, nn = nn, keep = keep, S = S,
+      basef = Ybase, W = Omega, Ht = Ht, sol = sol, nn = nn, keep = keep, S = S, type = type,
       settings = settings, b_pos = b_pos, bounds = bounds, nn_type = nn_type, v = v
     )
   }else if (type == "S") {