+++ R/reco.R
@@ -586,10 +586,39 @@ solveLin <- function(msx, mdx, verb = FALSE) {
   })
 }
 
-recoV <- function(basef, W, Ht, sol = "direct", nn = FALSE, nn_type = "osqp",
+# recoV <- function(basef, W, Ht, sol = "direct", nn = FALSE, nn_type = "osqp",
+#                   settings, b_pos = NULL, keep = "list", bounds = NULL, S, v){
+#   sol <- match.arg(sol, c("direct", "osqp"))
+#   nn_type <- match.arg(nn_type, c("osqp", "fbpp", "KAnn", "sntz"))
+#
+#   if(sol != "direct"){
+#     stop('Only sol = "direct" is available', call. = FALSE)
+#   }
+#
+#   if(max(v) > NCOL(basef)){
+#     stop('max(v) must be less or equal to ', NCOL(basef), call. = FALSE)
+#   }
+#
+#   Vcons <- .sparseDiagonal(NCOL(basef))[v, ,drop = FALSE]
+#   vcons <- basef[,v, drop = FALSE]
+#   Vt <- rbind(Ht, Vcons)
+#   vc <- cbind(Matrix(0, nrow = NROW(vcons), ncol = NROW(Ht)), vcons)
+#   switch(sol,
+#          direct = {
+#            out <- list()
+#
+#            lm_dx <- methods::as(t(vc) - Vt %*% t(basef), "CsparseMatrix")
+#            lm_sx <- Matrix::Matrix(Vt %*% W %*% t(Vt), sparse = TRUE, forceCheck = TRUE)
+#            out$recf <- t(t(basef) + W %*% t(Vt) %*% solveLin(lm_sx, lm_dx))
+#          }
+#   )
+#   return(out)
+# }
+
+recoV <- function(basef, W, Ht, sol = "direct", nn = FALSE, nn_type = "osqp", type = "M",
                   settings, b_pos = NULL, keep = "list", bounds = NULL, S, v){
   sol <- match.arg(sol, c("direct", "osqp"))
-  nn_type <- match.arg(nn_type, c("osqp", "fbpp", "KAnn", "sntz"))
+  nn_type <- match.arg(nn_type, c("osqp", "sntz"))
 
   if(sol != "direct"){
     stop('Only sol = "direct" is available', call. = FALSE)
@@ -603,13 +632,63 @@ recoV <- function(basef, W, Ht, sol = "direct", nn = FALSE, nn_type = "osqp",
   vcons <- basef[,v, drop = FALSE]
   Vt <- rbind(Ht, Vcons)
   vc <- cbind(Matrix(0, nrow = NROW(vcons), ncol = NROW(Ht)), vcons)
+
   switch(sol,
          direct = {
            out <- list()
+           if(type == "M"){
+             lm_dx <- methods::as(t(vc) - Vt %*% t(basef), "CsparseMatrix")
+             lm_sx <- Matrix::Matrix(Vt %*% W %*% t(Vt), sparse = TRUE, forceCheck = TRUE)
+             out$recf <- t(t(basef) + W %*% t(Vt) %*% solveLin(lm_sx, lm_dx))
+           }else{
+             stop("Type = S not yet available",call. = FALSE)
+           }
 
-           lm_dx <- methods::as(t(vc) - Vt %*% t(basef), "CsparseMatrix")
-           lm_sx <- Matrix::Matrix(Vt %*% W %*% t(Vt), sparse = TRUE, forceCheck = TRUE)
-           out$recf <- t(t(basef) + W %*% t(Vt) %*% solveLin(lm_sx, lm_dx))
+           if(nn & any(out$recf[,-v, drop = FALSE] < (-sqrt(.Machine$double.eps)))){
+             vbounds <- array(Inf, dim = list(NCOL(basef), 2, NROW(basef)))
+             vbounds[,1,] <- -vbounds[,1,]
+             vbounds[v,1,] <- vbounds[v,2,] <- t(basef[,v])
+
+             if(any(out$recf[,v, drop = FALSE] < (-sqrt(.Machine$double.eps)))){
+               warning("The immutable forecasts present negative values", call. = FALSE)
+             }
+
+             switch(nn_type,
+                    osqp = {
+                      if(isDiagonal(W)){
+                        P <- .sparseDiagonal(x = diag(W)^(-1))
+                      }else{
+                        R <- chol(W)
+                        P <- chol2inv(R)
+                      }
+                      id <- which(rowSums(out$recf<(-sqrt(.Machine$double.eps)))!=0)
+                      rec <- lapply(id, function(x){
+                        M_osqp(y = basef[x,], Ht = Ht, P = P, nn = nn,
+                               bounds = vbounds[,,x], settings = settings, b_pos = b_pos[!(b_pos %in% v)])
+                      })
+                      rec <- do.call("rbind",rec)
+                      out$recf[id,] <- do.call("rbind", rec[,"recf"])
+                      out$info <- do.call("rbind", rec[,"info"])
+                      colnames(out$info) <- c("obj_val", "run_time", "iter", "pri_res",
+                                              "status", "status_polish")
+                      rownames(out$info) <- id
+
+                      out$recf[-id,-v,drop=FALSE] <- out$recf[-id,-v,drop=FALSE] * (out$recf[-id,-v,drop=FALSE] > 0)
+                    },
+                    sntz = {
+                      na <- NROW(S)-NCOL(S)
+                      bottom <- out$recf[,-c(1:na),drop = FALSE]
+                      bottom[,-v, drop = FALSE] <- bottom[,-v, drop = FALSE]*(bottom[,-v, drop = FALSE] > 0)
+                      out$recf <- bottom %*% t(S)
+                    })
+           }else{
+             if(nn){
+               if(any(out$recf[,v, drop = FALSE] < (-sqrt(.Machine$double.eps)))){
+                 warning("The immutable forecasts present negative values", call. = FALSE)
+               }
+               out$recf[,-v, drop = FALSE] <- out$recf[,-v, drop = FALSE] * (out$recf[,-v, drop = FALSE] > 0)
+             }
+           }
          }
   )
   return(out)