+++ NEWS.md
@@ -1,9 +1,14 @@
-# FoReco 0.2.3
+# FoReco 0.2.4
 
+##### Major changes
+* Added `lcmat()` function.
+
+##### Minor changes
 * Fixed BU approach when the number of columns of basef is equal to the number of bottom time series `htsrec()`;
 * Fixed `score_index()`;
 * Fixed the `bounds` param when `type = "S"` in `htsrec()`, `thfrec()` and `octrec()`;
 * Add the possibility to fix base forecasts through the `v` param in `htsrec()`, `thfrec()` and `octrec()` - experimental;
+* Add two new type of optimal cross-temporal reconciliation (**cs_struc** and **t_struc**);
 * Improved docs and bug fixes.
 
 # FoReco 0.2.2